+++ src/components/form/time-picker.tsx
@@ -3,8 +3,6 @@ import { HTMLProps, forwardRef, useRef, useEffect, useState } from "react";
 import cx from "classnames";
 import { UseFormRegisterReturn, UseFormSetValue, FieldValues } from "react-hook-form";
 import IMask from 'imask';
-import { Button } from "../button";
-import { faTrashCan } from "@fortawesome/free-regular-svg-icons";
 
 type Props<T extends FieldValues> = UseFormRegisterReturn<string> &
   HTMLProps<HTMLInputElement> & {
@@ -13,11 +11,10 @@ type Props<T extends FieldValues> = UseFormRegisterReturn<string> &
     feedback?: string;
     className?: string;
     setValue: UseFormSetValue<T>;
-    delete: () => void;
   };
 
   export const TimePicker = forwardRef<HTMLInputElement, Props<FieldValues>>(
-    ({ labelText, hintText, feedback, className, setValue, delete: deleteTimePicker, ...rest }: Props<FieldValues>, ) => {
+    ({ labelText, hintText, feedback, className, setValue, ...rest }: Props<FieldValues>, ) => {
     const [hour, setHour] = useState<string>('');
     const [minute, setMinute] = useState<string>('');
     const [period, setPeriod] = useState<string>('AM');
@@ -140,14 +137,6 @@ type Props<T extends FieldValues> = UseFormRegisterReturn<string> &
           <option value="AM">AM</option>
           <option value="PM">PM</option>
         </select>
-        <Button
-            type="button"
-            variant="icon"
-            buttonText=""
-            className="ml-auto"
-            icon={faTrashCan}
-            onClick={() => deleteTimePicker()}
-          ></Button>
         </div>
         {feedback ? (
             <span className="text-red text-base leading-5">{feedback}</span>
