+++ src/routes/onboarding/set-medication-reminders/medication-section.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useForm } from "react-hook-form";
 import { Form, Button } from "../../../components"
 import { faTrashCan } from "@fortawesome/free-regular-svg-icons";
-import { MedicationTimeInput } from "./medication-time-input";
+import { ListTimeInput } from "./list-time-input";
 
 interface Props {
   index?: number;
@@ -11,14 +11,16 @@ interface Props {
 }
 
 export const MedicationSection: React.FC<Props> = ({ index, onDelete, elementId}) =>  {
+  const inputName = `medication${elementId}`
+
+  type TextInputValue = Record<typeof inputName, string>;
 
   const {
     register,
     unregister,
     formState: {errors},
-  } = useForm();
+  } = useForm<TextInputValue>();
 
-  const inputName = `medication${elementId}` || "medication6"
 
   return (
     <div className="flex flex-col gap-5 bg-white mx-auto min-w-[315px] max-w-[400px] px-4 py-3 w-full rounded-[16px]">
@@ -40,13 +42,13 @@ export const MedicationSection: React.FC<Props> = ({ index, onDelete, elementId}
           labelText="Medication name & amount"
           placeholder="Medication name & amount"
           type="text"
-          feedback={errors[inputName]?.message as string | undefined}
+          feedback={errors[inputName]?.message}
           {...register(inputName, {
             required: "This field is required",
           })}
           />
           <Form.AddMoreSection buttonText="+ Add more reminders">
-            <MedicationTimeInput />
+            <ListTimeInput label="Reminder" parentIndex={index} />
           </Form.AddMoreSection>
         </div>
   )
